RankService done

+++ accs/Models/AssignedReward.cs
@@ -7,7 +7,7 @@ namespace accs.Models
     {
         public int RewardId { get; set; }
 		[JsonIgnore] public virtual Reward Reward { get; set; }
-        public int UnitId { get; set; }
+        public ulong UnitId { get; set; }
         [JsonIgnore] public virtual Unit Unit { get; set; }
         public bool Display { get; set; }
